scripts/requirements-atlas.txt
requests
python-dotenv
psycopg2-binary
openai
anthropic
numpy
umap-learn
